Split vmecpp/__init__.py into core modules
__init__.py had grown to 2662 lines and held every Python-side definition, so
_continuation.py could only reach VmecInput/VmecOutput/run through
function-local `import vmecpp` calls, and _iteration.py could not annotate its
`vmec_input` parameter at all.

The definitions now live in core modules that __init__.py re-exports:

  _types.py        shared pydantic field types, enums, fixed array dimensions
  _input_files.py  INDATA vs. VMEC++ JSON detection and conversion
  _indata.py       VmecInput, set_profile
  _wout.py         VmecWOut, jxbout, mercier, threed1 tables
  _output.py       VmecOutput
  _run.py          run(), and the Fourier-continuation driver

The import graph is acyclic, so the lazy imports in _continuation.py are gone
and _iteration.py annotates `vmec_input: VmecInput`. What made the last cycle
unavoidable was that the continuation driver is a loop over run() calls, so
_run_fourier_continuation moved next to run(); _continuation.py keeps the pure
resolution-mapping machinery (interpolate_solution, _step_input).

No API change: __all__ is untouched and every name previously importable from
`vmecpp` still is (verified by diffing the imported surface -- names,
signatures and pydantic model fields -- before and after). The only names no
longer exposed at package level are private helpers and leaked stdlib imports;
the one that was used elsewhere, _progress_tip_shown, is set by __main__ on
vmecpp._run now.
