+++ src/vmecpp/__main__.py
@@ -7,6 +7,7 @@ import sys
 from pathlib import Path
 
 import vmecpp
+from vmecpp import _run
 
 
 def _parse_convert_arguments(argv: list[str]) -> argparse.Namespace:
@@ -89,7 +90,8 @@ def main() -> None:
         print(  # noqa: T201
             "Tip: Use the --legacy flag for classic table output."
         )
-        vmecpp._progress_tip_shown = True
+        # The tip is printed here already, so don't let run() print it again.
+        _run._progress_tip_shown = True
 
     vmec_input = vmecpp.VmecInput.from_file(args.input_file)
     output = vmecpp.run(vmec_input, max_threads=args.max_threads, verbose=verbose)
