+++ src/vmecpp/_input_files.py
@@ -1,0 +1,108 @@
+# SPDX-FileCopyrightText: 2024-present Proxima Fusion GmbH <info@proximafusion.com>
+#
+# SPDX-License-Identifier: MIT
+"""Detection of, and conversion between, the INDATA and VMEC++ JSON input formats."""
+
+from __future__ import annotations
+
+import contextlib
+import json
+import logging
+import os
+import tempfile
+from collections.abc import Generator
+from pathlib import Path
+
+from vmecpp import _util
+
+logger = logging.getLogger(__name__)
+
+
+def is_vmec2000_input(input_file: Path) -> bool:
+    """Returns true if the input file looks like a Fortran VMEC/VMEC2000 INDATA file."""
+    # we peek at the first few non-blank, non-comment lines in the file:
+    # if one of them is "&INDATA", then this is an INDATA file
+    with open(input_file) as f:
+        for line in f:
+            stripped_line = line.strip()
+            if not stripped_line or stripped_line.startswith("!"):
+                continue
+            return stripped_line == "&INDATA"
+    return False
+
+
+@contextlib.contextmanager
+def ensure_vmecpp_input(input_path: Path) -> Generator[Path, None, None]:
+    """If input_path looks like a Fortran INDATA file, convert it to a VMEC++ JSON input
+    and return the path to this new JSON file.
+
+    Otherwise assume it is a VMEC++ json input: simply return the input_path unchanged.
+    """
+    if is_vmec2000_input(input_path):
+        logger.debug(
+            f"VMEC++ is being run with input file '{input_path}', which looks like "
+            "a Fortran INDATA file. It will be converted to a VMEC++ JSON input "
+            "on the fly. Please consider permanently converting the input to a "
+            " VMEC++ input JSON using the //third_party/indata2json tool."
+        )
+
+        # We also add the PID to the output file to ensure that the output file
+        # is different for multiple processes that run indata_to_json
+        # concurrently on the same input, as it happens e.g. when the SIMSOPT
+        # wrapper is run under `mpirun`.
+        configuration_name = _util.get_vmec_configuration_name(input_path)
+        output_file = input_path.with_name(f"{configuration_name}.{os.getpid()}.json")
+
+        vmecpp_input_path = _util.indata_to_json(
+            input_path, output_override=output_file
+        )
+        assert vmecpp_input_path == output_file.resolve()
+        try:
+            yield vmecpp_input_path
+        finally:
+            os.remove(vmecpp_input_path)
+    else:
+        # if the file is not a VMEC2000 indata file, we assume
+        # it is a VMEC++ JSON input file
+        yield input_path
+
+
+@contextlib.contextmanager
+def ensure_vmec2000_input(input_path: Path) -> Generator[Path, None, None]:
+    """If input_path does not look like a VMEC2000 INDATA file, assume it is a VMEC++
+    JSON input file, convert it to VMEC2000's format and return the path to the
+    converted file.
+
+    Otherwise simply return the input_path unchanged.
+
+    Given a VMEC++ JSON input file with path 'path/to/[input.]NAME[.json]' the converted
+    INDATA file will have path 'some/tmp/dir/input.NAME'.
+    A temporary directory is used in order to avoid race conditions when calling this
+    function multiple times on the same input concurrently; the `NAME` section of the
+    file name is preserved as it is common to have logic that extracts it and re-uses
+    it e.g. to decide how related files should be called.
+    """
+
+    if is_vmec2000_input(input_path):
+        # nothing to do: must yield result on first generator call,
+        # then exit (via a return)
+        yield input_path
+        return
+
+    vmecpp_input_basename = input_path.name.removesuffix(".json").removeprefix("input.")
+    indata_file = f"input.{vmecpp_input_basename}"
+
+    with open(input_path) as vmecpp_json_f:
+        vmecpp_json_dict = json.load(vmecpp_json_f)
+
+    indata_contents = _util.vmecpp_json_to_indata(vmecpp_json_dict)
+
+    # Otherwise we actually need to perform the JSON -> INDATA conversion.
+    # We need the try/finally in order to correctly clean up after
+    # ourselves even in case of errors raised from the body of the `with`
+    # in user code.
+    with tempfile.TemporaryDirectory() as tmpdir:
+        out_path = Path(tmpdir) / indata_file
+        with open(out_path, "w") as out_f:
+            out_f.write(indata_contents)
+        yield out_path
